==> 02_script/00_importDatasets.Rmd <==
---
title: "PAN PEQUENOS FELINOS"
subtitle: "Ação 6.1."
author: 'Desenvolvimento: Fernando Lima, D.Sc.'
knit: (function(input, ...) {rmarkdown::render(input, output_dir=paste0(rprojroot::find_rstudio_root_file(), "/02_script/01_limpezaEOrganizacao/outputs/"))})
output:
  pdf_document:
---

## PREPARAÇÃO

Importação de dados (se necessário). Caso já tenha passado por esta etapa e não houve atualização nos arquivos vá para a próxima etapa.

### PACOTES

O script abaixo usa uma função para identificar se os pacotes necessários já estão instalados e os chama ou se é necessário a instalação.

```{r pacotes, message=FALSE, warning=FALSE}
rm(list = ls(all = TRUE))
source("funcoes/pacotes.R")
pacotes("here", "tidyverse", "httr") #suppdata para burlar captcha mas precisa de instalar RTools
```

### ATLANTIC CAMTRAPS

Dataset de dados registros de ocorrência de mamíferos de médio e grande porte da Mata Atlântica a partir de dados de armadilhas fotográficas:

Lima, F., et al. (2017), ATLANTIC-CAMTRAPS: a dataset of medium and large terrestrial mammal communities in the Atlantic Forest of South America. Ecology, 98: 2979-2979. https://doi.org/10.1002/ecy.1998

```{r}
# uma atualização da plataforma adicionou uma barreira contra automação de download. Se houver erro copiar o link e baixar o arquivo manualmente.

#download.file(url = "https://esajournals.onlinelibrary.wiley.com/action/downloadSupplement?doi=10.1002%2Fecy.1998&file=ecy1998-sup-0001-DataS1.zip",
#              destfile = "../01_input/data_papers/atlantic_camtraps.zip",
 #             mode = "wb"
  #            )

unzip(
    zipfile = "../01_input/data_papers/atlantic_camtraps.zip",
    exdir = "../01_input/data_papers"
    )
```

### ATLANTIC LARGE MAMMALS

Dataset de dados registros de ocorrência de mamíferos de médio e grande porte da Mata Atlântica a partir de dados de armadilhas fotográficas:

Souza Y., et al. 2019. ATLANTIC MAMMALS: a data set of assemblages of medium- and large-sized mammals of the Atlantic Forest of South America. Ecology 100(10):e02785. 10.1002/ecy.2785

```{r}
# uma atualização da plataforma adicionou uma barreira contra automação de download. Se houver erro copiar o link e baixar o arquivo manualmente.
#download.file(url = "https://esajournals.onlinelibrary.wiley.com/action/downloadSupplement?doi=10.1002%2Fecy.2785&file=ecy2785-sup-0001-DataS1.zip",
 #             destfile = "../01_input/data_papers/atlantic_large_mammals.zip",
  #            mode = "wb"
   #           )

unzip(
    zipfile = "../01_input/data_papers/atlantic_large_mammals.zip",
    exdir = "../01_input/data_papers"
    )
```

### NEOTROPICAL CARNIVORES



```{r}
# uma atualização da plataforma adicionou uma barreira contra automação de download. Se houver erro copiar o link e baixar o arquivo manualmente.
#download.file(url = "https://esajournals.onlinelibrary.wiley.com/action/downloadSupplement?doi=10.1002%2Fecy.3128&file=ecy3128-sup-0001-DataS1.zip",
 #             destfile = "../01_input/data_papers/neotropical_carnivores.zip",
  #            mode = "wb"
   #           )

unzip(
    zipfile = "../01_input/data_papers/neotropical_carnivores.zip",
    exdir = "../01_input/data_papers"
    )
```
