==> 02_script/01_prepareDatasets.Rmd <==
---
title: "PAN PEQUENOS FELINOS"
subtitle: "Ação 6.1."
author: 'Desenvolvimento: Fernando Lima, D.Sc.'
knit: (function(input, ...) {rmarkdown::render(input, output_dir=paste0(rprojroot::find_rstudio_root_file(), "/02_script/01_limpezaEOrganizacao/outputs/"))})
output:
  pdf_document:
---

Preparação de dados

## PREPARAÇÃO

### PACOTES

```{r pacotes, message=FALSE, warning=FALSE}
rm(list = ls(all = TRUE))
source("funcoes/pacotes.R")
pacotes("here", "tidyverse", "data.table", "sf", "terra", "fasterize", "tmap")
```

```{r}
# options
options(timeout = 1e6)
sf::sf_use_s2(FALSE)
```


### ATLANTIC CAMTRAPS

#### Importar "locations" 

```{r, message=FALSE}
# import sites
ca_si <- readr::read_csv(
    "../01_input/data_papers/ATLANTIC_CAMTRAPS_1-0_LOCATION.csv"
    ) %>%
    # select only id and coordinates
    dplyr::select(
        location_id, X, Y
        )
#ca_si
```

#### Importar "surveys"

```{r, message=FALSE}
# import sites names
ca_si_na <- readr::read_csv(
    "../01_input/data_papers/ATLANTIC_CAMTRAPS_1-0_SURVEY.csv"
    ) %>%
    # join sites names with coordinates
    dplyr::left_join(
        .,
        ca_si,
        by = "location_id"
        ) %>%
    # select only ids, year and coordinates
    dplyr::select(
        study_id, location_id, survey_id, yearfinish,  X, Y
        )
#ca_si_na
```

#### Importar "study"

```{r, message=FALSE}
# import study
ca_st_na <- readr::read_csv(
    "../01_input/data_papers/ATLANTIC_CAMTRAPS_1-0_STUDY.csv"
    ) %>% 
    dplyr::left_join(
        ., ca_si_na, by = "study_id"
        ) %>% 
    dplyr::select(
        location_id, survey_id, yearfinish,  X, Y, study
        )
#ca_st_na
```

#### Importar "species"

```{r, message=FALSE}
# import species names
ca_sp_na <- readr::read_csv(
    "../01_input/data_papers/ATLANTIC_CAMTRAPS_1-0_SPECIES.csv"
    ) %>%
    # remove unnecessary columns
    dplyr::select(
        order:genus, species_name, species_code
        )
#ca_sp_na
```

#### Importar "records"

```{r, message=FALSE}
# import records
ca_sp <- readr::read_csv(
    "../01_input/data_papers/ATLANTIC_CAMTRAPS_1-0_RECORDS.csv"
    ) %>% 
    # remove unnecessary columns
    dplyr::select(
        survey_id, species_code, presence_absence
        ) %>% 
    # remove absences
    dplyr::filter(
        presence_absence == 1
        ) %>% 
    # correction of Potus flavus
    dplyr::mutate(
        species_code = ifelse(
            species_code == "Potu_flav", "Poto_flav", species_code
            )
        ) %>% 
    # join with taxonomy information
    dplyr::left_join(
        ., ca_sp_na, by = "species_code"
        ) %>% 
    # remove unnacessary columns
    dplyr::select(
        -c(
            presence_absence,  species_code
            )
        )
#ca_sp
```

#### Integrar informações

```{r, message=FALSE}
# join spatial data
ca_occ <- ca_sp %>% 
    dplyr::left_join(
        ca_st_na, by = "survey_id"
        ) %>% 
    # rename columns
    dplyr::mutate(species = species_name,
                  longitude = as.numeric(X), 
                  latitude = as.numeric(Y), 
                  year = as.numeric(yearfinish),
                  Observation = "Camera Trapping",
                  SCALP = "C1",
                  source = "atlantic_camtrap") %>%
    #remove unnecessary columns
    dplyr::select(
        order:genus, species, longitude, latitude, year, source, Observation, SCALP, study
        ) %>% 
    # add id
    tibble::rowid_to_column()
#ca_occ
```

#### Exportar

```{r}
# export data
readr::write_csv(ca_occ, "data/atlantic_camtrap.csv")
```

### ATLANTIC LARGE MAMMALS

#### Importar dados

```{r, message=FALSE}
# import data
lm_occ <- readr::read_csv(
    "../01_input/data_papers/ATLANTIC_MAMMAL_MID_LARGE _assemblages_and_sites.csv"
    ) %>% 
    dplyr::mutate(
        genus = str_split(Actual_species_Name, " ", simplify = TRUE)[, 1],
        species = Actual_species_Name,
        longitude = as.numeric(Longitude), 
        latitude = as.numeric(Latitude), 
        year = as.numeric(Year_finish),
        Observation = Method,
        SCALP = "C1",
        study = Reference, 
        source = "atlantic_large_mammals"
        ) %>%
    dplyr::select(
        genus, species, longitude, latitude, year, source, Observation, SCALP, study
        ) %>% 
    tidyr::drop_na(
        longitude, latitude
        ) %>% 
    dplyr::mutate(
        species = str_trim(species)
        ) %>% 
    tibble::rowid_to_column() %>% 
    dplyr::left_join(
        .,
        rbind(
            unique(
                ca_occ[, c(2, 3, 4)]
                )
            ),
        by = "genus"
        ) %>% 
    dplyr::relocate(
        order:family,
        .before = 2
        )

lm_occ <- lm_occ[!grepl("sand_plot", lm_occ$Observation),]
lm_occ <- lm_occ[!grepl("mixed_method", lm_occ$Observation),]

sort(unique(lm_occ$Observation ))

#lm_occ
```

#### Exportar
```{r}
# export data
readr::write_csv(lm_occ, "data/atlantic_large_mammals.csv")
```

### NEOTROPICAL CARNIVORES

#### Importar dados
```{r,message=FALSE}
# import
neo_car_occ <- readr::read_csv("../01_input/data_papers/NEOTROPICAL_CARNIVORES_DATASET_2020-04.csv")
```

#### Filtrar registros duvidosos
```{r}
sort(unique(neo_car_occ$TYPE_REC))
neo_car_occ <- neo_car_occ[!grepl("Informal interview", neo_car_occ$TYPE_REC),]
neo_car_occ <- neo_car_occ[!grepl("Interview", neo_car_occ$TYPE_REC),]
neo_car_occ <- neo_car_occ[!grepl("Scratch", neo_car_occ$TYPE_REC),]
neo_car_occ <- neo_car_occ[!grepl("Sign", neo_car_occ$TYPE_REC),]
neo_car_occ <- neo_car_occ[!grepl("Scat", neo_car_occ$TYPE_REC),]
neo_car_occ <- neo_car_occ[!grepl("Vocalization", neo_car_occ$TYPE_REC),]
neo_car_occ <- neo_car_occ[!grepl("Track", neo_car_occ$TYPE_REC),]
neo_car_occ <- neo_car_occ[!grepl("Sighting", neo_car_occ$TYPE_REC),]
neo_car_occ <- neo_car_occ[!grepl("Literature review", neo_car_occ$TYPE_REC),]
neo_car_occ <- neo_car_occ[!grepl("Tissue", neo_car_occ$TYPE_REC),]

sort(unique(neo_car_occ$TYPE_REC))

#Precisa resolver o sinal em alguns anos
#RESOLVIDO
#sort(unique(neo_car_occ$COL_END_YR))

neo_car_occ <- neo_car_occ %>%
    dplyr::mutate(family = FAMILY,
                  genus = GENUS,
                  species = SPECIES,
                  name = SPECIES, 
                  longitude = LONG_X, 
                  latitude = LAT_Y,
                  COL_END_YR = gsub("<", "", COL_END_YR),
                  year = as.numeric(COL_END_YR),
                  Observation = TYPE_REC,
                  SCALP = "C1",
                  study = REFERENCE,
                  source = "neotropical_carnivores") %>%
    dplyr::select(
        family, genus, species, longitude, latitude, year, Observation, SCALP, source, study
        ) %>% 
    dplyr::mutate(
        species = str_trim(species)
        ) %>% 
    tibble::rowid_to_column() %>% 
    dplyr::left_join(
        .,
        unique(
            rbind(
                ca_occ[, c(2, 4)],
                lm_occ[, c(2, 4)]
                )
            ),
        by = "genus"
        ) %>% 
    dplyr::relocate(
        order, .after = 1
        )
#neo_car_occ
```

#### Exportar
```{r}
# export data
readr::write_csv(neo_car_occ, "data/neotropical_carnivores.csv")

```

## Tirelli et al. 2021

### Importar dados
```{r}
# import data
tirelli2021_occ <- readr::read_csv2(
    "../01_input/data_real_MunoasPampasCat.csv"
    ) %>%
  dplyr::mutate(
    species = "Leopardus munoai",
    genus = "Leopardus",
    family = "Felidae",
    order = "Carnivora",
    source = "tirelli2021",
    longitude = Long,
    latitude = Lat,
    year = as.numeric(Year),
    study = `Source/ Colectors`,
    Observation = Sample,
    SCALP = "C1"
    ) %>% 
  dplyr::relocate(
    order:family:genus:species,
    .before = 2,
    ) %>%
  dplyr::filter(
    !is.na(longitude)
    ) %>%
  dplyr::select(
    order, family, genus, species, longitude, latitude, year, source, Observation, SCALP, study
    ) %>%
  tibble::rowid_to_column(
    )

```

#### Exportar
```{r}
# export data
readr::write_csv(tirelli2021_occ, "data/tirelli2021.csv")

```


### GBIF

```{r}
# remove columns that only contain NA values
not_all_na <- function(x) any(!is.na(x))

sort(unique(gbif_lp$year))
sort(unique(gbif_lp$basisOfRecord))
sort(unique(gbif_lp$scientificName))

gbif_lp <- read.csv(
    "../01_input/GBIF_leopardus.csv", sep = "\t"
    ) %>%
    # filter data without coordinates
    dplyr::filter(
        !is.na(decimalLatitude)
    ) %>%
    dplyr::filter(
        !is.na(decimalLongitude)
    ) %>%
    dplyr::filter(
         year > 1998 #& basisOfRecord == "Occurrence"
    ) %>%
    dplyr::mutate(
        species = verbatimScientificName,
        longitude = decimalLongitude,
        latitude = decimalLatitude,
        source = "GBIF.org",
        study = "GBIF.org (16 December 2022) GBIF Occurrence Download https://doi.org/10.15468/dl.9cen59"
    ) #%>%
    dplyr::select(
        family, genus, species, longitude, latitude, year, source, study
    )
    
gbif_lp <- gbif_lp[grepl("Leopardus ", gbif_lp$species),]
```

### SALVE

```{r}
salve_wiedii <- readr::read_csv(
    "../01_input/salve-exportacao-ocorrencias-26-01-2026-10-33-30-181502271_lwiedii.csv"
    )

salve_guttulus <- readr::read_csv(
    "../01_input/salve-exportacao-ocorrencias-26-01-2026-16-58-04-1146751312_lguttulus.csv"
    )

salve_tigrinus <- readr::read_csv(
    "../01_input/salve-exportacao-ocorrencias-26-01-2026-17-02-13-493901872_ltigrinus.csv"
    )

salve_munoai <- readr::read_csv(
    "../01_input/salve-exportacao-ocorrencias-26-01-2026-17-04-20-1768857575_lmunoai.csv"
    )

salve_geoffroyi <- readr::read_csv(
    "../01_input/salve-exportacao-ocorrencias-26-01-2026-17-16-22-1563462613_lgeoffroyi.csv"
    )

salve_yagouaroundi <- readr::read_csv(
    "../01_input/salve-exportacao-ocorrencias-26-01-2026-17-18-37-109021950_hyagouaroundi.csv"
    )

salve_braccatus <- readr::read_csv(
    "../01_input/salve-exportacao-ocorrencias-27-01-2026-11-22-10-1331237110_lbraccatus.csv"
    )

salvePAN <- dplyr::bind_rows(
  salve_braccatus,
  salve_geoffroyi,
  salve_guttulus,
  salve_munoai,
  salve_tigrinus,
  salve_wiedii,
  salve_yagouaroundi
) %>%
  dplyr::mutate(
    source = "SALVE_202601",
    longitude = Longitude,
    latitude = Latitude,
    year = Ano,
    study = Referencia_bibliografica,
    Observation = Tipo_registro,
    SCALP = "C1",
    order = Ordem,
    family = Familia,
    genus = Genero,
    species = Especie
    ) %>%
  dplyr::select(
    order, family, genus, species, longitude, latitude, year, source, Observation, SCALP, study
    ) %>%
  tibble::rowid_to_column(
    )



```

#### Exportar
```{r}
# export data
readr::write_csv(salvePAN, "data/salvepanpf.csv")

```
