==> 02_script/02_load_Datasets.Rmd <==
---
title: "Organização e limpeza - Ação 6.1"
output: html_notebook
---

```{r}
library(tidyverse)
```


```{r}
# carregar dados do Atlantic Large Mammals
at_lar <- readr::read_csv("data/atlantic_large_mammals.csv")
at_lar

# Carregar dados do Atlantic Camtraps
at_cam <- readr::read_csv("data/atlantic_camtrap.csv")
at_cam

# Carregar dados do Neotropical Carnivores
neo_car <- readr::read_csv("data/neotropical_carnivores.csv")
neo_car

# Carregar dados do SALVE
salve_panpf <- readr::read_csv("data/salvepanpf.csv")
salve_panpf

# Carregar dados de Tirelli 2021
tirelli2021 <- readr::read_csv("data/tirelli2021.csv")
tirelli2021


# Junte os dados
occ <- dplyr::bind_rows(
  at_cam,
  at_lar,
  neo_car,
  salve_panpf,
  tirelli2021
  ) %>%
  # Selecione apenas os campos de interesse sem repetição
  dplyr::select(-1) %>%

  # marcar pontos onde colocolo se sobrepõe às espécies novas
  dplyr::group_by(longitude, latitude) %>%
  dplyr::filter(
    !(species == "Leopardus colocolo" &
        any(species %in% c("Leopardus munoai", "Leopardus braccatus")))
  ) %>%
  dplyr::ungroup() %>%
# reclassificar colocolo remanescente por latitude
  dplyr::mutate(
    species = dplyr::case_when(
      species == "Leopardus colocolo" & latitude > -27.000 ~ "Leopardus braccatus",
      species == "Leopardus colocolo" & latitude <= -27.000 ~ "Leopardus munoai",
      TRUE ~ species
    )
  ) %>%
  # garantia final de unicidade
  dplyr::distinct(species, longitude, latitude, .keep_all = TRUE)
```


```{r}
# vetor
occ_v <- occ %>% 
    tidyr::drop_na(longitude, latitude) %>% 
    dplyr::mutate(lon = longitude,
                  lat = latitude) %>% 
    dplyr::filter(lon > -180 & lon < 180) %>% 
    dplyr::filter(lat > -90 & lat < 90) %>% 
    sf::st_as_sf(coords = c("lon", "lat"), crs = 4326)
occ_v

plot(occ_v$geometry, pch = 20)

```

```{r}
mamiferos <- occ_v
mamiferos <- mamiferos %>%
  dplyr::filter(
    family == "Felidae"
  )

mamiferos <- mamiferos[!grepl("sp.", mamiferos$species),]

mamiferos <- mamiferos[!grepl("Panthera onca", mamiferos$species),]
mamiferos <- mamiferos[!grepl("Puma concolor", mamiferos$species),]
mamiferos <- mamiferos[!grepl("Felis catus", mamiferos$species),]
mamiferos <- mamiferos[!grepl("Lynx rufus", mamiferos$species),]
mamiferos <- mamiferos[!grepl("Leopardus guigna", mamiferos$species),]
mamiferos <- mamiferos[!grepl("Leopardus jacobita", mamiferos$species),]
mamiferos <- mamiferos[!grepl("Leopardus pardalis", mamiferos$species),]
mamiferos$species <- gsub("Puma yagouaroundi","Herpailurus yagouaroundi",
                          mamiferos$species)
mamiferos$genus <- gsub("Puma","Herpailurus",
                          mamiferos$genus)
sort(unique(occ_v$species))
sort(unique(mamiferos$species))
plot(mamiferos$geometry, pch = 20)

```

The code first modifies the "species" column, replacing "Puma yagouaroundi" with "Herpailurus yagouaroundi". Then, it adds the columns "order", "family", and "genus" with the specified values. Finally, it modifies the "Observation" column, replacing "cam_trap" with "Camera Trapping" and "Photo" with "Picture/Video". The resulting dataframe is filtered to include only rows where the species is "Herpailurus yagouaroundi" and the year is greater than 2002. The modified dataframe is stored back in the jaguarundi variable.

the case_when() function is used to specify the conditions for replacing values in the "Observation" column. The ~ symbol indicates the replacement value. The TRUE ~ Observation statement is used as the default case, which means that if none of the specified conditions match, the original value will be retained.
    

```{r}
jaguarundi <- occ_v

jaguarundi <- jaguarundi %>%
  mutate(
    species = gsub("Puma yagouaroundi", "Herpailurus yagouaroundi", species),
    order = "Carnivora",
    family = "Felidae",
    genus = "Herpailurus",
    Observation = gsub("cam_trap", "Camera Trapping", Observation),
    Observation = gsub("line_transect|mixed_method", "Direct Sighting", Observation),
    Observation = gsub("vehicles", "Roadkill", Observation),
    Observation = case_when(
      Observation %in% c("Photo", "Photo, Video", "Video") ~ "Picture/Video",
      TRUE ~ Observation
    ),
    SCALP = ifelse(
      Observation %in% c("Camera Trapping", "Capture", "Picture/Video"),
      "C1",
      ifelse(
        Observation %in% c("Carcass", "Roadkill", "Specimen"),
        "C2",
        ifelse(
          Observation == "Direct Sighting",
          "C3",
          Observation
        )
      )
    ),
    study = ifelse(is.na(study), "Unpublished data", study)
  ) %>%
  filter(
    species == "Herpailurus yagouaroundi",
    year > 2002
  )

sort(unique(jaguarundi$Observation))
sort(unique(jaguarundi$SCALP))
plot(jaguarundi$geometry, pch = 20)
```

```{r}
# export data
readr::write_csv(jaguarundi, "data/herpailurusYagouaroundi.csv")

```




```{r}
brasil <- sf::st_read("../01_input/geodata/vetor_EstadosBR_LLWGS84/EstadosBR_IBGE_LLWGS84.shp")

gatosBrasil <- mamiferos[brasil, ] %>% 
  tibble::rowid_to_column()

plot(gatosBrasil$geometry, pch = 20)

gatosBrasil$genus <- word(gatosBrasil$species,1)

unique(sort(gatosBrasil$species))

unique(sort(gatosBrasil$genus))

unique(sort(gatosBrasil$family))

unique(sort(gatosBrasil$order))

gatosBrasil <- gatosBrasil %>%
  filter(year>1999)

gatosBrasil %>%
  count(species)

```

```{r}
for (sp in unique(gatosBrasil$species)) {
  gatosBrasil %>%
    filter(species == sp) %>%
    write_csv(paste0("data/gatosBrasil_", sp, ".csv"))
}
```


```{r}
# export data
readr::write_csv(gatosBrasil, "data/gatosBrasil.csv")
```
